requirements.txt
tensorflow-gpu
scipy
matplotlib
Pillow
opencv-python
numpy
sklearn
scikit-image